🐛 Fixed a bug

+++ download.py
@@ -20,17 +20,19 @@ while Download == True:
             if lines:
                 url = lines[0]
             
-    os.system('youtube-dl ' + url)
-
-    if File == True:
-        with open(filename, 'r+') as f:
-            t = f.read()
-            to_delete = lines[0].strip()
-            f.seek(0)
-            for line in t.split('\n'):
-                if line != to_delete:
-                    f.write(line + '\n')
-            f.truncate()
+    if url and not url.__eq__('\n'):
+        os.system('youtube-dl ' + url)
+        if File == True:
+            with open(filename, 'r+') as f:
+                t = f.read()
+                to_delete = lines[0].strip()
+                f.seek(0)
+                for line in t.split('\n'):
+                    if line != to_delete:
+                        f.write(line + '\n')
+                f.truncate()
+        else:
+            Download = False
     else:
         Download = False
